requirements_v4.txt
streamlit>=1.28
pillow>=9
